+++ Src/Public/New-AsBuiltReport.ps1
@@ -70,7 +70,7 @@ function New-AsBuiltReport {
         Creates a VMware vSphere As Built Report in HTML format, using the configuration in the asbuiltreport.json file located in the C:\scripts\ folder.
         The report will be saved to c:\scripts.
     .NOTES
-        Version:        1.0.0
+        Version:        1.0.1
         Author(s):      Tim Carman / Matt Allford
         Twitter:        @tpcarman / @mattallford
         Github:         AsBuiltReport
@@ -83,7 +83,8 @@ function New-AsBuiltReport {
 
     #region Script Parameters
     [CmdletBinding(
-        PositionalBinding = $false
+        PositionalBinding = $false,
+        DefaultParameterSetName = 'Credential'
     )]
     param (
         [Parameter(
@@ -92,7 +93,7 @@ function New-AsBuiltReport {
             HelpMessage = 'Please specify which report type you wish to run.'
         )]
         [ValidateScript( {
-                $InstalledReportModules = Get-Module -Name "AsBuiltReport.*" -ListAvailable | Where-Object { $_.name -ne 'AsBuiltReport.Core' }
+                $InstalledReportModules = Get-Module -Name "AsBuiltReport.*" -ListAvailable | Where-Object { $_.name -ne 'AsBuiltReport.Core' } | Sort-Object -Property Version -Descending | Select-Object -Unique
                 $ValidReports = foreach ($InstalledReportModule in $InstalledReportModules) {
                     $NameArray = $InstalledReportModule.Name.Split('.')
                     "$($NameArray[-2]).$($NameArray[-1])"
@@ -113,7 +114,6 @@ function New-AsBuiltReport {
         [ValidateNotNullOrEmpty()]
         [Alias('Cluster', 'Server', 'IP')]
         [String[]] $Target,
-
         [Parameter(
             Position = 2,
             Mandatory = $true,
@@ -213,6 +213,10 @@ function New-AsBuiltReport {
             $Credential = New-Object System.Management.Automation.PSCredential ($Username, $SecurePassword)
         }
 
+        if (!(Test-Path $OutputPath)) {
+            Write-Error "OutputPath $OutputPath is not a valid directory path"
+            break
+        }
         #region Variable config
 
         # Import the AsBuiltReport JSON configuration file
@@ -281,12 +285,13 @@ function New-AsBuiltReport {
 
         #region Email Server Authentication
         # If Email Server Authentication is required, prompt user for credentials
-        if ($AsBuiltConfig.Email.Credentials) {
+        if ($SendEmail -and $AsBuiltConfig.Email.Credentials) {
             Clear-Host
             Write-Host '---------------------------------------------' -ForegroundColor Cyan
             Write-Host '  <        Email Server Credentials       >  ' -ForegroundColor Cyan
             Write-Host '---------------------------------------------' -ForegroundColor Cyan
             $MailCredentials = Get-Credential -Message "Please enter the credentials for $($AsBuiltConfig.Email.Server)"
+            Clear-Host
         }
         #endregion Email Server Authentication
 
@@ -357,7 +362,7 @@ Register-ArgumentCompleter -CommandName 'New-AsBuiltReport' -ParameterName 'Repo
         $fakeBoundParameter
     )
 
-    $InstalledReportModules = Get-Module -Name "AsBuiltReport.*" -ListAvailable | Where-Object { $_.name -ne 'AsBuiltReport.Core' }
+    $InstalledReportModules = Get-Module -Name "AsBuiltReport.*" -ListAvailable | Where-Object { $_.name -ne 'AsBuiltReport.Core' } | Sort-Object -Property Version -Descending | Select-Object -Unique
     $ValidReports = foreach ($InstalledReportModule in $InstalledReportModules) {
         $NameArray = $InstalledReportModule.Name.Split('.')
         "$($NameArray[-2]).$($NameArray[-1])"