+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # AsBuiltReport.Core Changelog
 
+## [1.0.2] - 2019-06-21
+### Changed
+- Update module manifest icon and release notes URI
+
+## [1.0.1] - 2019-05-17
+### Added
+- Add DefaultParameterSetName to `New-AsBuiltReport` so a user can run the cmdlet without any parameters
+- Check if `OutputPath` parameter for `New-AsBuiltReport` exists before running the report script and error if it doesn't exist
+
+### Changed
+- Fix issues when multiple versions of the same AsBuiltReport report module is installed
+
 ## [1.0.0] - 2019-03-26
 ### Added
 - `AsBuiltReport.Core` module created to provide core as built report functionality